# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your super ApplicationAPI",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/register/training-center": {
            "post": {
                "tags": [
                    "Sign up"
                ],
                "description": "Sign up as Training center",
                "operationId": "trainingCenterRegister",
                "requestBody": {
                    "description": "Pass user credentials",
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password",
                                    "first_name",
                                    "last_name",
                                    "phone"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "center123@mail.ru"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "center"
                                    },
                                    "first_name": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "First name"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Last name"
                                    },
                                    "phone": {
                                        "type": "number",
                                        "example": "+374 98-066-083"
                                    },
                                    "tax_identity_number": {
                                        "type": "string",
                                        "example": "777777"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry, wrong email address or password. Please try again"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/create/trainer": {
            "post": {
                "tags": [
                    "Training center actions"
                ],
                "description": "Sign up as trainer",
                "operationId": "TrainingCenterCreateTrainer",
                "requestBody": {
                    "description": "Pass user credentials",
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password",
                                    "first_name",
                                    "last_name",
                                    "phone"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "trainer123@mail.ru"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "trainer"
                                    },
                                    "first_name": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "First name"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Last name"
                                    },
                                    "phone": {
                                        "type": "number",
                                        "example": "+374 98-066-083"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry, wrong email address or password. Please try again"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/delete/trainer/{id}": {
            "delete": {
                "tags": [
                    "Training center actions"
                ],
                "description": "Delete trainer",
                "operationId": "Delete trainer",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id trainer",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Empty list"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/update/trainer/{id}": {
            "post": {
                "tags": [
                    "Training center actions"
                ],
                "description": "Update trainer",
                "operationId": "Update trainer",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id trainer",
                        "required": true
                    }
                ],
                "requestBody": {
                    "description": "Pass user credentials",
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "trainer123@mail.ru"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "trainer"
                                    },
                                    "first_name": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "First name"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Last name"
                                    },
                                    "phone": {
                                        "type": "number",
                                        "example": "+374 98-066-083"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Empty list"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/trainers/list": {
            "post": {
                "tags": [
                    "Training center actions"
                ],
                "description": "Sign up as trainer",
                "operationId": "Trainers list",
                "responses": {
                    "204": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Empty list"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/trainer/{id}": {
            "post": {
                "tags": [
                    "Training center actions"
                ],
                "description": "Sign up as trainer",
                "operationId": "Current trainer",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id trainer",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Empty list"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        }
    },
    "components": {
...
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}